requirements.txt
black==22.1.0
click==8.0.4
Django==2.2
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
